requirements.txt
httpx
psycopg2-binary
python-dotenv
